+++ CHANGELOG.md
@@ -1,1 +1,4 @@
-Fixed error when downloading a single log file.
+* You can now export the opened log as a CSV file!
+* A progress bar is now shown when downloading logs over SFTP.
+* The field count shown in the side bar no longer includes individual array items.
+* The name of ".rlog" files registered with the OS is now "Robot Logs" instead of "Robot logs" (to better match other formats).